Improvements to RustaCUDA after dogfooding it in my path tracer.
