+++ rustacuda_core/README.md
@@ -1,0 +1,5 @@
+Minimal kernel-support crate for [RustaCUDA](https://github.com/bheisler/RustaCUDA).
+
+This crate contains a small number of types which can be helpful when compiling Rust code to PTX
+for use with RustaCUDA. It is not a general-purpose kernel support crate, and it is not intended to
+be used in a standalone manner.
